app/requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-filter==0.10.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
-e git+https://github.com/yigor/django-jsonfield.git#egg=django-jsonfield
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
